gemini1.5/requirements.txt
# Project dependencies
google-genai
pandas
python-dotenv
bibtexparser
rispy # For reading RIS files
# PyPDF2 # Optional: uncomment if PDF parsing is needed 